requirements.txt
Flask==3.1.0
gunicorn==22.0.0
pillow==11.2.1
python-dotenv==1.2.1
requests==2.32.5
Werkzeug==3.1.3
